+++ smartcontract/programs/doublezero-serviceability/src/processors/user/mod.rs
@@ -7,4 +7,5 @@ pub mod create_subscribe;
 pub mod delete;
 pub mod reject;
 pub mod requestban;
+pub mod resource_onchain_helpers;
 pub mod update;